Revise how unsupported browsers are handled
Browsers other than IE will be allowed to load the GUI.
However, if there is an error, the error boundary will display an unsupported browser message recommending they update to the latest version of a supported browser.

For the purposes of this PR, ‘supported’ means you can load the gui (anything other than IE), and recommended is just the browsers that meet our minimum requirements for Chrome, Firefox, Edge and Safari.