Refactor step and step image naming conventions
Step Ids and step file names did not follow a consistent pattern. To make it easier to automate in the future, things were renamed in the following ways:

StepId: camel case, prefix with the short name that we usually use to refer to the tutorial, e.g. name, music, chaseGame. Followed by a name for the step, e.g. nameAddSprite, nameAddBackdrop. When there are two of a step, add ‘2’ to the end e.g. nameAddSprite2

file-naming: all step files should be in the steps directory. The name is the ‘kebab’ case of the stepId, followed by the language-code, then the file extension. For example the english version of the file for `nameAddSprite` would be `name-add-sprite.en.gif`, `chaseGameAddSprite2` would be `chase-game-add-sprite2.en.gif`