Allow Android to request a thumbnail generated on load
Usually we generate a thumbnail when a project is changed. If a project is imported on Android it doesn’t have a thumbnail, and if you just run it without changes it still doesn’t have a thumbnail.

Adding `saveThumbnailOnLoad` flag that Android sets if the project doesn’t have a thumbnail.