build: don't confuse npm install with unpublished version numbers
Even with `--offline` and installing a dependency that's contained in an
internal worksace, `npm --workspace a install b` will put the latest
published version of `b` (that it knows about) into the lock file rather
than the current monorepo version of `b`.

The only way I've found to get the version updated correctly in both
`package.json` and `package-lock.json` is the way that
@MiroslavDionisiev did it in the previous versioning script: just
manually hammer the version number, then get `npm i --package-lock-only`
to make the lock file consistent afterward.